Reorganize state management
