+++ src/tui/contexts/ErrorContext.tsx
@@ -1,11 +1,13 @@
 import React, { useCallback, useContext, useState } from "react";
-import { Text } from "ink";
-import { useLoaderContext } from "./LoaderContext";
+import { Box, Text } from "ink";
 import { useLogContext } from "./LogContext";
+import OptionList from "../components/OptionList";
+import { useAppStateContext } from "./AppStateContext";
 
 export type ThrowError = (err: string | null) => void;
 
 export interface IErrorContext {
+  errorThrown: boolean;
   throwError: ThrowError;
   clearError: () => void;
 }
@@ -13,21 +15,41 @@ export interface IErrorContext {
 export const ErrorContext = React.createContext<IErrorContext | undefined>(undefined);
 
 export const useErrorContext = () => {
-  return useContext(ErrorContext) as IErrorContext;
+  const context = useContext(ErrorContext);
+  if (!context) {
+    throw new Error("useErrorContext must be used within a ErrorContextProvider");
+  }
+  return context;
 };
 
-const ErrorFallback: React.FC<{ error: string }> = ({ error }) => {
+const ErrorFallback: React.FC<{ error: string; onClearError: () => void }> = ({
+  error,
+  onClearError,
+}) => {
   return (
-    <Text>
-      Something went wrong:
-      <Text> {error}</Text>
-    </Text>
+    <Box>
+      <Box>
+        <Text>
+          Something went wrong:
+          <Text> {error}</Text>
+        </Text>
+      </Box>
+
+      <OptionList
+        options={{
+          ["Clear Error"]: {
+            label: "Clear Error",
+            onSelect: onClearError,
+          },
+        }}
+      />
+    </Box>
   );
 };
 
 export const ErrorContextProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
   const { clearLog } = useLogContext();
-  const { setIsLoading } = useLoaderContext();
+  const { setIsLoading } = useAppStateContext();
 
   const [err, setErr] = useState<string | null>(null);
 
@@ -41,15 +63,18 @@ export const ErrorContextProvider: React.FC<{ children: React.ReactNode }> = ({ 
   );
 
   const clearError = useCallback(() => throwError(""), [throwError]);
+  const errorThrown = err !== null && err.length > 0;
 
   return (
     <ErrorContext.Provider
       value={{
+        errorThrown,
         throwError,
         clearError,
       }}
     >
-      {err !== null && err.length > 0 ? <ErrorFallback error={err} /> : children}
+      {errorThrown && <ErrorFallback error={err} onClearError={clearError} />}
+      {children}
     </ErrorContext.Provider>
   );
 };
