+++ src/tui/index.tsx
@@ -3,12 +3,13 @@ import { render, Box } from "ink";
 
 import { ErrorContextProvider } from "./contexts/ErrorContext";
 import { LogContextProvider } from "./contexts/LogContext";
-import { LoaderContextProvider } from "./contexts/LoaderContext";
 import { ConfigContextProvider } from "./contexts/ConfigContext";
-import { AppContextProvider } from "./contexts/AppContext";
+import { AppActionContextProvider } from "./contexts/AppActionContext";
 import { LayoutWrapper } from "./layouts/Layout";
 import { LAYOUT_KEY } from "./layouts/keys";
 import App from "./App";
+import { DownloadContextProvider } from "./contexts/DownloadContext";
+import { AppStateContextProvider } from "./contexts/AppStateContext";
 
 export default function renderTUI() {
   const clearANSI: string = process.platform === "win32" ? "u001b[H\u001bc" : "\u001b[2J";
@@ -18,20 +19,22 @@ export default function renderTUI() {
   process.stdout.write(clearANSI);
 
   render(
-    <LogContextProvider>
-      <LoaderContextProvider>
-        <ErrorContextProvider>
-          <ConfigContextProvider>
-            <LayoutWrapper initialLayout={LAYOUT_KEY.SEARCH_LAYOUT}>
-              <AppContextProvider>
-                <Box width={80} marginLeft={1} paddingRight={4}>
-                  <App />
-                </Box>
-              </AppContextProvider>
-            </LayoutWrapper>
-          </ConfigContextProvider>
-        </ErrorContextProvider>
-      </LoaderContextProvider>
-    </LogContextProvider>
+    <AppStateContextProvider>
+      <LogContextProvider>
+        <LayoutWrapper initialLayout={LAYOUT_KEY.SEARCH_LAYOUT}>
+          <ErrorContextProvider>
+            <ConfigContextProvider>
+              <DownloadContextProvider>
+                <AppActionContextProvider>
+                  <Box width={80} marginLeft={1} paddingRight={4}>
+                    <App />
+                  </Box>
+                </AppActionContextProvider>
+              </DownloadContextProvider>
+            </ConfigContextProvider>
+          </ErrorContextProvider>
+        </LayoutWrapper>
+      </LogContextProvider>
+    </AppStateContextProvider>
   );
 }
