+++ src/tui/layouts/search/search-filter/FilterInput.tsx
@@ -2,14 +2,14 @@ import React, { useCallback } from "react";
 import { useFocus } from "ink";
 
 import Input from "../../../components/Input";
-import { useAppContext } from "../../../contexts/AppContext";
 import { FilterKey } from "./Filter.data";
+import { useAppStateContext } from "../../../contexts/AppStateContext";
 
 const FilterInput: React.FC<{
   label: string;
   filterKey: FilterKey;
 }> = ({ label, filterKey }) => {
-  const { filters, setFilters } = useAppContext();
+  const { filters, setFilters } = useAppStateContext();
 
   const { isFocused } = useFocus({ autoFocus: true });
